requirements.txt
fastapi
uvicorn
httpx
pydantic
jinja2
python-dotenv
slowapi
pydantic[email]
